refactor: merge copy.ts paste.ts cut.ts into one file
